Clean up typer code a bit by splitting into separate files
